adding menu for subtasks

+++ app/src/main/java/com/example/todoapp/DashboardActivity.kt
@@ -107,8 +107,16 @@ class DashboardActivity : AppCompatActivity() {
                             activity.updateToDo(list[position])
                         }
                         R.id.menu_delete->{
-                            activity.dbHandler.deleteToDo(list[position].id)
-                            activity.refreshList()
+                            val dialog = AlertDialog.Builder(activity)
+                            dialog.setTitle("Are you sure?")
+                            dialog.setMessage("Do you want to delete this task?")
+                            dialog.setPositiveButton("Continue") { _: DialogInterface, _: Int ->
+                                activity.dbHandler.deleteToDo(list[position].id)
+                                activity.refreshList()
+                            }
+                            dialog.setNegativeButton("Cancel") { _: DialogInterface, _: Int ->
+                            }
+                            dialog.show()
                         }
                         R.id.menu_mark_as_completed->{
                             activity.dbHandler.updateToDoItemCompletedStatus(list[position].id,true)
